Implement binary parsing source generators
Implements binary parsing source generators powered by attribute level metadata. Additional metadata is discovered through data types. This paves the way for implementing a source generator for automatically generating documentation for the binary representation

+++ src/BglReader/Airport/Subsections/Types/ComType.cs
@@ -1,6 +1,6 @@
 ﻿namespace BglReader.Airport.Subsections.Types;
 
-public enum ComType : ushort
+public enum ComType : byte
 {
     Atis = 0x0001,
     Multicom = 0x0002,
